Turn Parameters.DEBUG to false in NALTest
abf84d0c79df30d52c1a843f4b7a090e0b8ffb5e saw the introduction of

```
Parameters.DEBUG = true;
```

In NALTest which then turned on a bunch of verbose logging elsewhere in
the code base.  I switched this to false as it seems unnecessary for
normal test runs.  Alternatively it might be good to change this to be :

```
Parameters.DEBUG = Boolean.getBoolean("nar.test.debug");
```

But this also has drawbacks as you'd need to ensure that runs before all
other tests, etc...  This is the reason I didn't do it this way.  Simply
changing the value to false seemed the most straightforward way.

Addresses opennars/opennars#132